+++ tencentcloud/ess/v20201111/models.go
@@ -434,6 +434,7 @@ type CreateFlowByFilesRequestParams struct {
 	// TEXT - 内容文本控件
 	// MULTI_LINE_TEXT - 多行文本控件
 	// CHECK_BOX - 勾选框控件
+	// ATTACHMENT - 附件
 	// 注：默认字体大小为 字号12
 	Components []*Component `json:"Components,omitempty" name:"Components"`
 
@@ -487,6 +488,7 @@ type CreateFlowByFilesRequest struct {
 	// TEXT - 内容文本控件
 	// MULTI_LINE_TEXT - 多行文本控件
 	// CHECK_BOX - 勾选框控件
+	// ATTACHMENT - 附件
 	// 注：默认字体大小为 字号12
 	Components []*Component `json:"Components,omitempty" name:"Components"`