+++ tencentcloud/organization/v20210331/models.go
@@ -571,6 +571,10 @@ type OrgMember struct {
 	// 安全信息绑定状态  未绑定：Unbound，待激活：Valid，绑定成功：Success，绑定失败：Failed
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	BindStatus *string `json:"BindStatus,omitempty" name:"BindStatus"`
+
+	// 成员权限状态 已确认：Confirmed ，待确认：UnConfirmed
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	PermissionStatus *string `json:"PermissionStatus,omitempty" name:"PermissionStatus"`
 }
 
 type OrgPermission struct {