+++ tencentcloud/tcex/v20200727/client.go
@@ -63,6 +63,10 @@ func NewDescribeInvocationResultResponse() (response *DescribeInvocationResultRe
 }
 
 // DescribeInvocationResult
+// 产品控制台已经下线
+//
+// 
+//
 // 获取服务调用结果。和InvokeService接口配置合适，其InvokeId参数为InvokeService接口返回的RequestId。
 //
 // 可能返回的错误码:
@@ -77,6 +81,10 @@ func (c *Client) DescribeInvocationResult(request *DescribeInvocationResultReque
 }
 
 // DescribeInvocationResult
+// 产品控制台已经下线
+//
+// 
+//
 // 获取服务调用结果。和InvokeService接口配置合适，其InvokeId参数为InvokeService接口返回的RequestId。
 //
 // 可能返回的错误码:
@@ -120,6 +128,10 @@ func NewInvokeServiceResponse() (response *InvokeServiceResponse) {
 }
 
 // InvokeService
+// 产品控制台已经下线
+//
+// 
+//
 // 通过传入文档url，测试服务算法。此接口需要和DescribeInvocationResult接口配置使用，该接口使用InvokeService返回的RequestId作为InvokeId参数，用于查询调用结果。
 //
 // 可能返回的错误码:
@@ -138,6 +150,10 @@ func (c *Client) InvokeService(request *InvokeServiceRequest) (response *InvokeS
 }
 
 // InvokeService
+// 产品控制台已经下线
+//
+// 
+//
 // 通过传入文档url，测试服务算法。此接口需要和DescribeInvocationResult接口配置使用，该接口使用InvokeService返回的RequestId作为InvokeId参数，用于查询调用结果。
 //
 // 可能返回的错误码: